3.4:
Fix/Update the "Import From Twitch" button: Due to Twitch's new API limitations, from now on this function will only be able to read and import your own user's followed channels. It will not be possible to import the list of other user's followed channels. For 99% of users, this will not be a significant change.
Update Twitch OAuth permissions accordingly
You will need to disconnect from twitch (in the Options menu) and re-authenticate for these changes to take effect
Bring back the functionality of following the stream you are watching with the click of a button